put builder in its own mod, separate tests and benchmarks
