+++ examples/threadsafe.rs
@@ -1,12 +1,11 @@
-#![feature(scoped)]
 extern crate coroutine;
 extern crate num_cpus;
 
-use std::thread;
-
-use coroutine::coroutine::{State, Coroutine};
-
+#[cfg(feature = "enable-clonable-handle")]
 fn main() {
+    use std::thread;
+
+    use coroutine::coroutine::{State, Coroutine};
 
     let mut threads = Vec::with_capacity(num_cpus::get());
 
@@ -21,7 +20,7 @@ fn main() {
 
     for i in 0..num_cpus::get() {
         let coro = coro.clone();
-        let t = thread::Builder::new().name(format!("Thread {}", i)).scoped(move|| {
+        let t = thread::Builder::new().name(format!("Thread {}", i)).spawn(move|| {
             loop {
                 match coro.state() {
                     State::Finished | State::Panicked => break,
@@ -37,4 +36,9 @@ fn main() {
     for t in threads.into_iter() {
         t.join();
     }
+}
+
+#[cfg(not(feature = "enable-clonable-handle"))]
+fn main() {
+    panic!("Require `enable-clonable-handle` feature");
 }
