+++ src/coroutine_unique.rs
@@ -84,6 +84,7 @@ use std::any::Any;
 use std::cell::UnsafeCell;
 use std::ops::Deref;
 use std::ptr::Unique;
+use std::fmt::{self, Debug};
 
 use context::Context;
 use stack::{StackPool, Stack};
@@ -484,213 +485,5 @@ impl Environment {
         COROUTINE_ENVIRONMENT.with(|env| unsafe {
             &mut *env.get()
         })
-    }
-}
-
-/// Coroutine configuration. Provides detailed control over the properties and behavior of new Coroutines.
-///
-/// ```ignore
-/// let coro = Builder::new().name(format!("Coroutine #{}", 1))
-///                          .stack_size(4096)
-///                          .spawn(|| println!("Hello world!!"));
-///
-/// coro.resume().unwrap();
-/// ```
-pub struct Builder {
-    opts: Options,
-}
-
-impl Builder {
-    /// Generate the base configuration for spawning a Coroutine, from which configuration methods can be chained.
-    pub fn new() -> Builder {
-        Builder {
-            opts: Default::default(),
-        }
-    }
-
-    /// Name the Coroutine-to-be. Currently the name is used for identification only in panic messages.
-    pub fn name(mut self, name: String) -> Builder {
-        self.opts.name = Some(name);
-        self
-    }
-
-    /// Set the size of the stack for the new Coroutine.
-    pub fn stack_size(mut self, size: usize) -> Builder {
-        self.opts.stack_size = size;
-        self
-    }
-
-    /// Spawn a new Coroutine, and return a handle for it.
-    pub fn spawn<F>(self, f: F) -> Handle
-        where F: FnOnce() + Send + 'static
-    {
-        Coroutine::spawn_opts(f, self.opts)
-    }
-}
-
-/// Spawn a new Coroutine
-///
-/// Equavalent to `Coroutine::spawn`.
-pub fn spawn<F>(f: F) -> Handle
-    where F: FnOnce() + Send + 'static
-{
-    Builder::new().spawn(f)
-}
-
-/// Get the current Coroutine
-///
-/// Equavalent to `Coroutine::current`.
-pub fn current() -> &'static Handle {
-    Coroutine::current()
-}
-
-/// Resume a Coroutine
-///
-/// Equavalent to `Coroutine::resume`.
-pub fn resume(coro: &Handle) -> ResumeResult<()> {
-    coro.resume()
-}
-
-/// Yield the current Coroutine
-///
-/// Equavalent to `Coroutine::sched`.
-pub fn sched() {
-    Coroutine::sched()
-}
-
-#[cfg(test)]
-mod test {
-    use std::sync::mpsc::channel;
-
-    use super::Coroutine;
-    use super::Builder;
-
-    #[test]
-    fn test_coroutine_basic() {
-        let (tx, rx) = channel();
-        Coroutine::spawn(move|| {
-            tx.send(1).unwrap();
-        }).resume().ok().expect("Failed to resume");
-
-        assert_eq!(rx.recv().unwrap(), 1);
-    }
-
-    #[test]
-    fn test_coroutine_yield() {
-        let (tx, rx) = channel();
-        let coro = Coroutine::spawn(move|| {
-            tx.send(1).unwrap();
-
-            Coroutine::sched();
-
-            tx.send(2).unwrap();
-        });
-        coro.resume().ok().expect("Failed to resume");
-        assert_eq!(rx.recv().unwrap(), 1);
-        assert!(rx.try_recv().is_err());
-
-        coro.resume().ok().expect("Failed to resume");
-
-        assert_eq!(rx.recv().unwrap(), 2);
-    }
-
-    #[test]
-    fn test_coroutine_spawn_inside() {
-        let (tx, rx) = channel();
-        Coroutine::spawn(move|| {
-            tx.send(1).unwrap();
-
-            Coroutine::spawn(move|| {
-                tx.send(2).unwrap();
-            }).join().ok().expect("Failed to join");
-
-        }).join().ok().expect("Failed to join");
-
-        assert_eq!(rx.recv().unwrap(), 1);
-        assert_eq!(rx.recv().unwrap(), 2);
-    }
-
-    #[test]
-    fn test_coroutine_panic() {
-        let coro = Coroutine::spawn(move|| {
-            panic!("Panic inside a coroutine!!");
-        });
-        assert!(coro.join().is_err());
-    }
-
-    #[test]
-    fn test_coroutine_child_panic() {
-        Coroutine::spawn(move|| {
-            let _ = Coroutine::spawn(move|| {
-                panic!("Panic inside a coroutine's child!!");
-            }).join();
-        }).join().ok().expect("Failed to join");
-    }
-
-    #[test]
-    fn test_coroutine_resume_after_finished() {
-        let coro = Coroutine::spawn(move|| {});
-
-        // It is already finished, but we try to resume it
-        // Idealy it would come back immediately
-        assert!(coro.resume().is_ok());
-
-        // Again?
-        assert!(coro.resume().is_ok());
-    }
-
-    #[test]
-    fn test_coroutine_resume_itself() {
-        let coro = Coroutine::spawn(move|| {
-            // Resume itself
-            Coroutine::current().resume().ok().expect("Failed to resume");
-        });
-
-        assert!(coro.resume().is_ok());
-    }
-
-    #[test]
-    fn test_coroutine_yield_in_main() {
-        Coroutine::sched();
-    }
-
-    #[test]
-    fn test_builder_basic() {
-        let (tx, rx) = channel();
-        Builder::new().name("Test builder".to_string()).spawn(move|| {
-            tx.send(1).unwrap();
-        }).join().ok().expect("Failed to join");
-        assert_eq!(rx.recv().unwrap(), 1);
-    }
-}
-
-#[cfg(test)]
-mod bench {
-    use test::Bencher;
-
-    use super::Coroutine;
-
-    #[bench]
-    fn bench_coroutine_spawning(b: &mut Bencher) {
-        b.iter(|| {
-            let _ = Coroutine::spawn(move|| {});
-        });
-    }
-
-    #[bench]
-    fn bench_context_switch(b: &mut Bencher) {
-        let coro = Coroutine::spawn(|| {
-            loop {
-                // 3. Save current context
-                // 4. Switch
-                Coroutine::sched();
-            }
-        });
-
-        b.iter(|| {
-            // 1. Save current context
-            // 2. Switch
-            let _ = coro.resume();
-        });
     }
 }
