+++ src/lib.rs
@@ -30,8 +30,8 @@ pub use coroutine_clonable as coroutine;
 #[cfg(not(feature = "enable-clonable-handle"))]
 pub use coroutine_unique as coroutine;
 
-pub use coroutine::Builder;
-pub use coroutine::{Coroutine, Handle, spawn, sched, current};
+pub use builder::Builder;
+pub use coroutine::{Coroutine, Handle, ResumeResult};
 
 mod context;
 #[cfg(feature = "enable-clonable-handle")]
@@ -39,6 +39,42 @@ pub mod coroutine_clonable;
 #[cfg(not(feature = "enable-clonable-handle"))]
 pub mod coroutine_unique;
 
+pub mod builder;
 mod stack;
 mod thunk; // use self-maintained thunk, because std::thunk is temporary. May be replaced by FnBox in the future.
 mod sys;
+
+#[cfg(test)]
+mod tests;
+#[cfg(test)]
+mod benchmarks;
+
+/// Spawn a new Coroutine
+///
+/// Equavalent to `Coroutine::spawn`.
+pub fn spawn<F>(f: F) -> Handle
+    where F: FnOnce() + Send + 'static
+{
+    Builder::new().spawn(f)
+}
+
+/// Get the current Coroutine
+///
+/// Equavalent to `Coroutine::current`.
+pub fn current() -> &'static Handle {
+    Coroutine::current()
+}
+
+/// Resume a Coroutine
+///
+/// Equavalent to `Coroutine::resume`.
+pub fn resume(coro: &Handle) -> ResumeResult<()> {
+    coro.resume()
+}
+
+/// Yield the current Coroutine
+///
+/// Equavalent to `Coroutine::sched`.
+pub fn sched() {
+    Coroutine::sched()
+}
